:wheelchair: Make default outputPath as process.cwd()

+++ src/build.js
@@ -35,7 +35,8 @@ module.exports = function (webpackConfig) {
     }, global.vusionConfig.webpack);
 
     // Remove output directory and copy assets
-    shell.rm('-rf', webpackConfig.output.path);
+    if (webpackConfig.output.path !== process.cwd())
+        shell.rm('-rf', webpackConfig.output.path);
     if (global.vusionConfig.assetsPath)
         shell.cp('-r', path.resolve(process.cwd(), global.vusionConfig.assetsPath), webpackConfig.output.path);
